fix: Display image
Image rendering fix. For some reason, whether it was the way I was doing the course or something else, the product images weren't showing.
At the end of Product CRUD, all the methods were tried where during the 'Update' method the images were added to '/images/product', but they were not saved as an 'ImageURL' field and thus were not displayed.